分离gen-arch gen-design
